refactor(next-list-item): drop NextList/NextListDivider container, rename snippet
NextList/NextListDivider were byte-identical to List/ListDivider (same
--list-item-border-radius var), so ship no new names for them. Only the items
differ (pressed-scale recipe + interaction context) and keep the Next name.

- remove NextListRoot (package) and NextList/NextListDivider (snippet); reuse
  List/ListDivider from ui:list as container/divider
- rename ui:next-list snippet -> ui:next-list-item (file, registry id, import
  paths), matching the next-list-item CSS recipe; it now exports only the
  NextList*Item components
- rename docs/examples/react/next-list -> next-list-item; update ComponentExample
  names, Storybook stories, and stackflow activities/snippet to the List container
- list.mdx composition uses <List.Root> with <NextList.Item> slots
