+++ docs/examples/react/next-list-item/check.tsx
@@ -1,10 +1,11 @@
 import { Badge, HStack } from "@seed-design/react";
-import { NextList, NextListDivider, NextListCheckItem } from "seed-design/ui/next-list";
+import { List, ListDivider } from "seed-design/ui/list";
+import { NextListCheckItem } from "seed-design/ui/next-list-item";
 import { Checkmark } from "seed-design/ui/checkbox";
 
 export default function ListCheckbox() {
   return (
-    <NextList as="fieldset" width="360px">
+    <List as="fieldset" width="360px">
       <NextListCheckItem
         title={
           <HStack gap="x1_5">
@@ -16,18 +17,18 @@ export default function ListCheckbox() {
         suffix={<Checkmark tone="neutral" size="large" />}
         defaultChecked
       />
-      <NextListDivider as="div" />
+      <ListDivider as="div" />
       <NextListCheckItem
         prefix={<Checkmark tone="neutral" size="large" />}
         title="마케팅 정보 수신 동의"
         detail="마케팅 정보를 받으시겠습니까?"
         defaultChecked
       />
-      <NextListDivider as="div" />
+      <ListDivider as="div" />
       <NextListCheckItem
         prefix={<Checkmark tone="neutral" size="large" variant="ghost" />}
         title="Ghost Variant"
       />
-    </NextList>
+    </List>
   );
 }
