+++ docs/examples/react/next-list-item/clickable.tsx
@@ -8,7 +8,8 @@ import {
 } from "@karrotmarket/react-monochrome-icon";
 import { PrefixIcon, Icon, Box } from "@seed-design/react";
 import { useCallback, useState } from "react";
-import { NextList, NextListDivider, NextListItem, NextListButtonItem, NextListLinkItem } from "seed-design/ui/next-list";
+import { List, ListDivider } from "seed-design/ui/list";
+import { NextListItem, NextListButtonItem, NextListLinkItem } from "seed-design/ui/next-list-item";
 import { ActionButton } from "seed-design/ui/action-button";
 import { ToggleButton } from "seed-design/ui/toggle-button";
 
@@ -26,7 +27,7 @@ export default function ListClickable() {
   }, []);
 
   return (
-    <NextList width="full">
+    <List width="full">
       <NextListItem
         title="NextListItem은 클릭할 수 없어요. 눌러보세요."
         detail="우측의 Action Button만 클릭할 수 있어요"
@@ -42,7 +43,7 @@ export default function ListClickable() {
           </ActionButton>
         }
       />
-      <NextListDivider />
+      <ListDivider />
       <NextListButtonItem
         title="NextListButtonItem은 클릭할 수 있어요. 눌러보세요."
         detail="리스트 항목 전체와 우측의 Toggle Button 각각을 클릭할 수 있어요"
@@ -57,7 +58,7 @@ export default function ListClickable() {
           </>
         }
       />
-      <NextListDivider />
+      <ListDivider />
       <NextListButtonItem
         title="NextListButtonItem은 클릭할 수 있어요. 눌러보세요."
         detail="리스트 항목 전체와 우측의 커스텀 버튼 각각을 클릭할 수 있어요"
@@ -77,7 +78,7 @@ export default function ListClickable() {
           </>
         }
       />
-      <NextListDivider />
+      <ListDivider />
       <NextListLinkItem
         title="NextListLinkItem도 클릭할 수 있어요. 눌러보세요."
         detail="리스트 항목 전체와 우측의 Action Button 각각을 클릭할 수 있어요"
@@ -94,6 +95,6 @@ export default function ListClickable() {
         target="_blank"
         rel="noreferrer"
       />
-    </NextList>
+    </List>
   );
 }
