+++ docs/examples/react/next-list-item/preview.tsx
@@ -1,4 +1,5 @@
-import { NextList, NextListDivider, NextListItem } from "seed-design/ui/next-list";
+import { List, ListDivider } from "seed-design/ui/list";
+import { NextListItem } from "seed-design/ui/next-list-item";
 import { ListHeader } from "seed-design/ui/list-header";
 import {
   IconILowercaseSerifCircleLine,
@@ -10,16 +11,16 @@ export default function ListPreview() {
   return (
     <VStack width="360px">
       <ListHeader as="h2">리스트 헤더</ListHeader>
-      <NextList width="full">
+      <List width="full">
         <NextListItem title="기본 리스트 아이템" />
-        <NextListDivider />
+        <ListDivider />
         <NextListItem
           prefix={<Icon svg={<IconPersonCircleLine />} />}
           title="아이콘이 있는 리스트 아이템"
           detail="부가 정보가 포함된 설명"
           suffix={<Icon svg={<IconILowercaseSerifCircleLine />} />}
         />
-      </NextList>
+      </List>
     </VStack>
   );
 }
