+++ docs/examples/react/next-list-item/radio.tsx
@@ -1,10 +1,11 @@
 import { RadioGroup } from "@seed-design/react/primitive";
-import { NextList, NextListDivider, NextListRadioItem } from "seed-design/ui/next-list";
+import { List, ListDivider } from "seed-design/ui/list";
+import { NextListRadioItem } from "seed-design/ui/next-list-item";
 import { Radiomark } from "seed-design/ui/radio-group";
 
 export default function ListRadio() {
   return (
-    <NextList width="360px" asChild>
+    <List width="360px" asChild>
       <RadioGroup.Root defaultValue="option1" aria-label="옵션 선택">
         <NextListRadioItem
           value="option1"
@@ -12,14 +13,14 @@ export default function ListRadio() {
           detail="첫 번째 선택지"
           suffix={<Radiomark tone="neutral" size="large" />}
         />
-        <NextListDivider as="div" />
+        <ListDivider as="div" />
         <NextListRadioItem
           prefix={<Radiomark tone="neutral" size="large" />}
           value="option2"
           title="옵션 2"
           detail="두 번째 선택지"
         />
-        <NextListDivider as="div" />
+        <ListDivider as="div" />
         <NextListRadioItem
           prefix={<Radiomark tone="neutral" size="large" />}
           value="option3"
@@ -27,6 +28,6 @@ export default function ListRadio() {
           detail="세 번째 선택지"
         />
       </RadioGroup.Root>
-    </NextList>
+    </List>
   );
 }
