+++ docs/examples/react/next-list-item/switch.tsx
@@ -1,18 +1,19 @@
 import { IconTrashcanLine } from "@karrotmarket/react-monochrome-icon";
 import { IconSparkle2 } from "@karrotmarket/react-multicolor-icon";
 import { Icon } from "@seed-design/react";
-import { NextList, NextListDivider, NextListSwitchItem } from "seed-design/ui/next-list";
+import { List, ListDivider } from "seed-design/ui/list";
+import { NextListSwitchItem } from "seed-design/ui/next-list-item";
 import { Switchmark } from "seed-design/ui/switch";
 
 export default function ListSwitch() {
   return (
-    <NextList width="360px">
+    <List width="360px">
       <NextListSwitchItem
         title="삭제하기 전에 확인"
         prefix={<Icon svg={<IconTrashcanLine />} />}
         suffix={<Switchmark tone="neutral" />}
       />
-      <NextListDivider />
+      <ListDivider />
       <NextListSwitchItem
         title="메시지 요약"
         detail="핵심 내용만 빠르게 확인해보세요."
@@ -20,6 +21,6 @@ export default function ListSwitch() {
         suffix={<Switchmark tone="neutral" />}
         defaultChecked
       />
-    </NextList>
+    </List>
   );
 }
