+++ docs/public/__docs__/index.json
@@ -1221,7 +1221,7 @@
             {
               "id": "list",
               "title": "List",
-              "description": "가로 행으로 구성된 콘텐츠를 표현하는 컴포넌트입니다. 행의 동작과 pressed scale 같은 상호작용은 NextList item이 책임집니다.",
+              "description": "가로 행으로 구성된 콘텐츠를 표현하는 컴포넌트입니다.",
               "docUrl": "/react/components/list",
               "snippetKey": "react/ui:list",
               "snippets": [
