+++ docs/public/__registry__/react/ui/index.json
@@ -680,14 +680,14 @@
     {
       "snippets": [
         {
-          "path": "next-list.tsx",
+          "path": "next-list-item.tsx",
           "dependencies": {
             "@seed-design/react": "^2.1.0",
             "@seed-design/css": "^2.2.0"
           }
         }
       ],
-      "id": "next-list",
+      "id": "next-list-item",
       "dependencies": [
         "@seed-design/react"
       ]
