+++ docs/public/__registry__/react/ui/next-list-item.json
@@ -1,0 +1,16 @@
+{
+  "id": "next-list-item",
+  "dependencies": [
+    "@seed-design/react"
+  ],
+  "snippets": [
+    {
+      "path": "next-list-item.tsx",
+      "content": "/**\n * @file ui:next-list-item\n * @requires @seed-design/react@^2.1.0\n * @requires @seed-design/css@^2.2.0\n **/\n\n\"use client\";\n\nimport { NextList as SeedNextList } from \"@seed-design/react\";\nimport * as React from \"react\";\n\nexport interface NextListItemProps\n  extends Omit<SeedNextList.ItemProps, \"title\" | \"prefix\" | \"children\"> {\n  title: React.ReactNode;\n  detail?: React.ReactNode;\n  prefix?: React.ReactNode;\n  suffix?: React.ReactNode;\n}\n\n/**\n * @see https://seed-design.io/react/components/list\n */\nexport const NextListItem = React.forwardRef<HTMLLIElement, NextListItemProps>(\n  ({ title, detail, prefix, suffix, ...otherProps }, ref) => {\n    return (\n      <SeedNextList.Item ref={ref} {...otherProps}>\n        {prefix && <SeedNextList.Prefix>{prefix}</SeedNextList.Prefix>}\n        <SeedNextList.Content>\n          <SeedNextList.Title>{title}</SeedNextList.Title>\n          {detail && <SeedNextList.Detail>{detail}</SeedNextList.Detail>}\n        </SeedNextList.Content>\n        {suffix && <SeedNextList.Suffix>{suffix}</SeedNextList.Suffix>}\n      </SeedNextList.Item>\n    );\n  },\n);\nNextListItem.displayName = \"NextListItem\";\n\nexport interface NextListButtonItemProps\n  extends Omit<SeedNextList.ButtonItemProps, \"title\" | \"prefix\" | \"children\"> {\n  title: React.ReactNode;\n  detail?: React.ReactNode;\n  prefix?: React.ReactNode;\n  suffix?: React.ReactNode;\n}\n\n/**\n * @see https://seed-design.io/react/components/list\n */\nexport const NextListButtonItem = React.forwardRef<HTMLButtonElement, NextListButtonItemProps>(\n  ({ title, detail, prefix, suffix, ...otherProps }, ref) => {\n    return (\n      <SeedNextList.ButtonItem ref={ref} {...otherProps}>\n        {prefix && <SeedNextList.Prefix>{prefix}</SeedNextList.Prefix>}\n        <SeedNextList.Content>\n          <SeedNextList.Title>{title}</SeedNextList.Title>\n          {detail && <SeedNextList.Detail>{detail}</SeedNextList.Detail>}\n        </SeedNextList.Content>\n        {suffix && <SeedNextList.Suffix>{suffix}</SeedNextList.Suffix>}\n      </SeedNextList.ButtonItem>\n    );\n  },\n);\nNextListButtonItem.displayName = \"NextListButtonItem\";\n\nexport interface NextListLinkItemProps\n  extends Omit<SeedNextList.AnchorItemProps, \"title\" | \"prefix\" | \"children\"> {\n  title: React.ReactNode;\n  detail?: React.ReactNode;\n  prefix?: React.ReactNode;\n  suffix?: React.ReactNode;\n}\n\n/**\n * @see https://seed-design.io/react/components/list\n */\nexport const NextListLinkItem = React.forwardRef<HTMLAnchorElement, NextListLinkItemProps>(\n  ({ title, detail, prefix, suffix, ...otherProps }, ref) => {\n    return (\n      <SeedNextList.AnchorItem ref={ref} {...otherProps}>\n        {prefix && <SeedNextList.Prefix>{prefix}</SeedNextList.Prefix>}\n        <SeedNextList.Content>\n          <SeedNextList.Title>{title}</SeedNextList.Title>\n          {detail && <SeedNextList.Detail>{detail}</SeedNextList.Detail>}\n        </SeedNextList.Content>\n        {suffix && <SeedNextList.Suffix>{suffix}</SeedNextList.Suffix>}\n      </SeedNextList.AnchorItem>\n    );\n  },\n);\nNextListLinkItem.displayName = \"NextListLinkItem\";\n\nexport interface NextListCheckItemProps\n  extends Omit<SeedNextList.CheckboxItemProps, \"title\" | \"prefix\" | \"children\"> {\n  title: React.ReactNode;\n  detail?: React.ReactNode;\n  prefix?: React.ReactNode;\n  suffix?: React.ReactNode;\n}\n\n/**\n * @see https://seed-design.io/react/components/list\n */\nexport const NextListCheckItem = React.forwardRef<HTMLInputElement, NextListCheckItemProps>(\n  ({ title, detail, prefix, suffix, ...otherProps }, ref) => {\n    return (\n      <SeedNextList.CheckboxItem ref={ref} {...otherProps}>\n        {prefix && <SeedNextList.Prefix>{prefix}</SeedNextList.Prefix>}\n        <SeedNextList.Content>\n          <SeedNextList.Title>{title}</SeedNextList.Title>\n          {detail && <SeedNextList.Detail>{detail}</SeedNextList.Detail>}\n        </SeedNextList.Content>\n        {suffix && <SeedNextList.Suffix>{suffix}</SeedNextList.Suffix>}\n      </SeedNextList.CheckboxItem>\n    );\n  },\n);\nNextListCheckItem.displayName = \"NextListCheckItem\";\n\nexport interface NextListRadioItemProps\n  extends Omit<SeedNextList.RadioItemProps, \"title\" | \"prefix\" | \"children\"> {\n  title: React.ReactNode;\n  detail?: React.ReactNode;\n  prefix?: React.ReactNode;\n  suffix?: React.ReactNode;\n}\n\n/**\n * @see https://seed-design.io/react/components/list\n */\nexport const NextListRadioItem = React.forwardRef<HTMLInputElement, NextListRadioItemProps>(\n  ({ title, detail, prefix, suffix, ...otherProps }, ref) => {\n    return (\n      <SeedNextList.RadioItem ref={ref} {...otherProps}>\n        {prefix && <SeedNextList.Prefix>{prefix}</SeedNextList.Prefix>}\n        <SeedNextList.Content>\n          <SeedNextList.Title>{title}</SeedNextList.Title>\n          {detail && <SeedNextList.Detail>{detail}</SeedNextList.Detail>}\n        </SeedNextList.Content>\n        {suffix && <SeedNextList.Suffix>{suffix}</SeedNextList.Suffix>}\n      </SeedNextList.RadioItem>\n    );\n  },\n);\nNextListRadioItem.displayName = \"NextListRadioItem\";\n\nexport interface NextListSwitchItemProps\n  extends Omit<SeedNextList.SwitchItemProps, \"title\" | \"prefix\" | \"children\"> {\n  title: React.ReactNode;\n  detail?: React.ReactNode;\n  prefix?: React.ReactNode;\n  suffix?: React.ReactNode;\n}\n\n/**\n * @see https://seed-design.io/react/components/list\n */\nexport const NextListSwitchItem = React.forwardRef<HTMLInputElement, NextListSwitchItemProps>(\n  ({ title, detail, prefix, suffix, ...otherProps }, ref) => {\n    return (\n      <SeedNextList.SwitchItem ref={ref} {...otherProps}>\n        {prefix && <SeedNextList.Prefix>{prefix}</SeedNextList.Prefix>}\n        <SeedNextList.Content>\n          <SeedNextList.Title>{title}</SeedNextList.Title>\n          {detail && <SeedNextList.Detail>{detail}</SeedNextList.Detail>}\n        </SeedNextList.Content>\n        {suffix && <SeedNextList.Suffix>{suffix}</SeedNextList.Suffix>}\n      </SeedNextList.SwitchItem>\n    );\n  },\n);\nNextListSwitchItem.displayName = \"NextListSwitchItem\";\n\n/**\n * This file is a snippet from SEED Design, helping you get started quickly with @seed-design/* packages.\n * You can extend this snippet however you want.\n */\n",
+      "dependencies": {
+        "@seed-design/react": "^2.1.0",
+        "@seed-design/css": "^2.2.0"
+      }
+    }
+  ]
+}
