+++ docs/registry/react/ui/next-list-item.tsx
@@ -1,18 +1,7 @@
 "use client";
 
-import {
-  NextList as SeedNextList,
-  Divider as SeedDivider,
-  type DividerProps as SeedDividerProps,
-} from "@seed-design/react";
+import { NextList as SeedNextList } from "@seed-design/react";
 import * as React from "react";
-
-export interface NextListProps extends SeedNextList.RootProps {}
-
-/**
- * @see https://seed-design.io/react/components/list
- */
-export const NextList = SeedNextList.Root;
 
 export interface NextListItemProps
   extends Omit<SeedNextList.ItemProps, "title" | "prefix" | "children"> {
@@ -175,32 +164,3 @@ export const NextListSwitchItem = React.forwardRef<HTMLInputElement, NextListSwi
   },
 );
 NextListSwitchItem.displayName = "NextListSwitchItem";
-
-export interface NextListDividerProps extends SeedDividerProps {
-  /**
-   * @default "li"
-   */
-  as?: SeedDividerProps["as"];
-
-  /**
-   * @default true
-   */
-  "aria-hidden"?: SeedDividerProps["aria-hidden"];
-}
-
-/**
- * @see https://seed-design.io/react/components/list
- */
-export const NextListDivider = React.forwardRef<HTMLLIElement, NextListDividerProps>(
-  ({ as = "li", "aria-hidden": ariaHidden = true, ...props }, ref) => {
-    return (
-      <SeedDivider
-        as={as}
-        aria-hidden={ariaHidden}
-        ref={ref as React.ForwardedRef<HTMLHRElement>}
-        {...props}
-      />
-    );
-  },
-);
-NextListDivider.displayName = "NextListDivider";
