+++ examples/stackflow-spa/src/activities/ActivityNextListButtonItem.tsx
@@ -20,7 +20,8 @@ import { AppScreen, AppScreenContent } from "seed-design/ui/app-screen";
 import { ActionButton } from "seed-design/ui/action-button";
 import { Avatar } from "seed-design/ui/avatar";
 import { IdentityPlaceholder } from "seed-design/ui/identity-placeholder";
-import { NextList, NextListDivider, NextListButtonItem } from "seed-design/ui/next-list";
+import { List, ListDivider } from "seed-design/ui/list";
+import { NextListButtonItem } from "seed-design/ui/next-list-item";
 
 const prefixVariants = [
   { key: "none", element: null },
@@ -104,7 +105,7 @@ const ActivityNextListButtonItem: StaticActivityComponentType<
           await new Promise((resolve) => setTimeout(resolve, 1000));
         }}
       >
-        <NextList>
+        <List>
           {prefixVariants.map((prefix) =>
             contentVariants.map((content) =>
               suffixVariants.map((suffix, suffixIndex) => {
@@ -127,13 +128,13 @@ const ActivityNextListButtonItem: StaticActivityComponentType<
                       disabled={content.disabled}
                       onClick={() => window.alert(`Row clicked: ${key}`)}
                     />
-                    {showDivider && <NextListDivider />}
+                    {showDivider && <ListDivider />}
                   </Fragment>
                 );
               }),
             ),
           )}
-        </NextList>
+        </List>
       </AppScreenContent>
     </AppScreen>
   );
