+++ examples/stackflow-spa/src/activities/ActivityNextListItem.tsx
@@ -19,7 +19,8 @@ import { AppScreen, AppScreenContent } from "seed-design/ui/app-screen";
 import { ActionButton } from "seed-design/ui/action-button";
 import { Avatar } from "seed-design/ui/avatar";
 import { IdentityPlaceholder } from "seed-design/ui/identity-placeholder";
-import { NextList, NextListDivider, NextListItem } from "seed-design/ui/next-list";
+import { List, ListDivider } from "seed-design/ui/list";
+import { NextListItem } from "seed-design/ui/next-list-item";
 import { Fragment } from "react";
 
 const prefixVariants = [
@@ -99,7 +100,7 @@ const ActivityNextListItem: StaticActivityComponentType<"ActivityNextListItem"> 
           await new Promise((resolve) => setTimeout(resolve, 1000));
         }}
       >
-        <NextList>
+        <List>
           {prefixVariants.map((prefix) =>
             contentVariants.map((content) =>
               suffixVariants.map((suffix, suffixIndex) => {
@@ -120,13 +121,13 @@ const ActivityNextListItem: StaticActivityComponentType<"ActivityNextListItem"> 
                       suffix={suffix.element}
                       highlighted={content.highlighted}
                     />
-                    {showDivider && <NextListDivider />}
+                    {showDivider && <ListDivider />}
                   </Fragment>
                 );
               }),
             ),
           )}
-        </NextList>
+        </List>
       </AppScreenContent>
     </AppScreen>
   );
