+++ examples/stackflow-spa/src/activities/ActivityNextListSwitchItem.tsx
@@ -10,7 +10,8 @@ import {
   AppBarRight,
 } from "seed-design/ui/app-bar";
 import { AppScreen, AppScreenContent } from "seed-design/ui/app-screen";
-import { NextList, NextListDivider, NextListSwitchItem } from "seed-design/ui/next-list";
+import { List, ListDivider } from "seed-design/ui/list";
+import { NextListSwitchItem } from "seed-design/ui/next-list-item";
 import { Icon } from "@seed-design/react";
 import { IdentityPlaceholder } from "seed-design/ui/identity-placeholder";
 import { Avatar } from "seed-design/ui/avatar";
@@ -90,7 +91,7 @@ const ActivityNextListSwitchItem: StaticActivityComponentType<
           await new Promise((resolve) => setTimeout(resolve, 1000));
         }}
       >
-        <NextList>
+        <List>
           {positionVariants.map((position) =>
             contentVariants.map((content) =>
               stateVariants.map((state) => {
@@ -122,7 +123,7 @@ const ActivityNextListSwitchItem: StaticActivityComponentType<
                             disabled={state.disabled}
                             suffix={suffix.element}
                           />
-                          {showDivider && <NextListDivider as="div" />}
+                          {showDivider && <ListDivider as="div" />}
                         </Fragment>
                       );
                     });
@@ -155,7 +156,7 @@ const ActivityNextListSwitchItem: StaticActivityComponentType<
                             disabled={state.disabled}
                             prefix={prefix.element}
                           />
-                          {showDivider && <NextListDivider as="div" />}
+                          {showDivider && <ListDivider as="div" />}
                         </Fragment>
                       );
                     });
@@ -167,7 +168,7 @@ const ActivityNextListSwitchItem: StaticActivityComponentType<
               }),
             ),
           )}
-        </NextList>
+        </List>
       </AppScreenContent>
     </AppScreen>
   );
