+++ packages/react/src/components/NextList/NextList.namespace.ts
@@ -7,7 +7,6 @@ export {
   NextListItem as Item,
   NextListPrefix as Prefix,
   NextListRadioItem as RadioItem,
-  NextListRoot as Root,
   NextListSuffix as Suffix,
   NextListSwitchItem as SwitchItem,
   NextListTitle as Title,
@@ -19,7 +18,6 @@ export {
   type NextListItemProps as ItemProps,
   type NextListPrefixProps as PrefixProps,
   type NextListRadioItemProps as RadioItemProps,
-  type NextListRootProps as RootProps,
   type NextListSuffixProps as SuffixProps,
   type NextListSwitchItemProps as SwitchItemProps,
   type NextListTitleProps as TitleProps,
