+++ packages/react/src/components/NextList/index.ts
@@ -7,7 +7,6 @@ export {
   NextListItem,
   NextListPrefix,
   NextListRadioItem,
-  NextListRoot,
   NextListSuffix,
   NextListSwitchItem,
   NextListTitle,
@@ -19,7 +18,6 @@ export {
   type NextListItemProps,
   type NextListPrefixProps,
   type NextListRadioItemProps,
-  type NextListRootProps,
   type NextListSuffixProps,
   type NextListSwitchItemProps,
   type NextListTitleProps,
